bug: fix by adding check for db in each method (#8)

+++ adapter.js
@@ -45,18 +45,27 @@ export function adapter(env, Datastore) {
       return Promise.resolve({ ok: true });
     },
     createDocument: async ({ db, id, doc }) => {
+      if (!db) {
+        return Promise.reject({ ok: false, msg: "database not found!" });
+      }
       db = new Datastore({ filename: dbFullname(db) });
       doc._id = id || cuid();
       const result = await db.insert(doc);
       return Promise.resolve({ ok: equals(result, doc), id: result._id });
     },
     retrieveDocument: async ({ db, id }) => {
+      if (!db) {
+        return Promise.reject({ ok: false, msg: "database not found!" });
+      }
       db = new Datastore({ filename: dbFullname(db) });
       const doc = await db.findOne({ _id: id });
       // swap ids
       return Promise.resolve(compose(omit(["_id"]), assoc("id", doc._id))(doc));
     },
     updateDocument: async ({ db, id, doc }) => {
+      if (!db) {
+        return Promise.reject({ ok: false, msg: "database not found!" });
+      }
       db = new Datastore({ filename: dbFullname(db) });
       // swap ids
       doc = toInternalId(doc);
@@ -64,6 +73,9 @@ export function adapter(env, Datastore) {
       return Promise.resolve({ ok: true });
     },
     removeDocument: async ({ db, id }) => {
+      if (!db) {
+        return Promise.reject({ ok: false, msg: "database not found!" });
+      }
       db = new Datastore({ filename: dbFullname(db) });
       const result = await db.removeOne({ _id: id });
       if (!result) return Promise.resolve({ ok: false, message: "not found" });
@@ -71,6 +83,9 @@ export function adapter(env, Datastore) {
     },
     listDocuments: async (d) => {
       let { db } = d;
+      if (!db) {
+        return Promise.reject({ ok: false, msg: "database not found!" });
+      }
       db = new Datastore({ filename: dbFullname(db) });
       let results = await db.find();
 
@@ -92,17 +107,26 @@ export function adapter(env, Datastore) {
       return Promise.resolve({ ok: true, docs: results });
     },
     queryDocuments: async ({ db, query }) => {
+      if (!db) {
+        return Promise.reject({ ok: false, msg: "database not found!" });
+      }
       db = new Datastore({ filename: dbFullname(db) });
       const results = await db.find(query.selector);
       return Promise.resolve({ ok: true, docs: results });
     },
     indexDocuments: ({ db, name, fields }) => {
+      if (!db) {
+        return Promise.reject({ ok: false, msg: "database not found!" });
+      }
       // noop - db is not built for
       // optimizability yet! will add this when
       // supported
       return Promise.resolve({ ok: true });
     },
     bulkDocuments: ({ db, docs }) => {
+      if (!db) {
+        return Promise.reject({ ok: false, msg: "database not found!" });
+      }
       db = new Datastore({ filename: dbFullname(db) });
       return bulk({ db, docs })
         .map((results) => ({ ok: true, results }))
